error handeling

+++ Application/Acitivities/Delete.cs
@@ -4,12 +4,13 @@ using System.Threading.Tasks;
 using Domain;
 using MediatR;
 using Persistence;
+using Application.Core;
 
 namespace Application.Acitivities
 {
     public class Delete
     {
-        public class Command: IRequest
+        public class Command: IRequest<Result<Unit>>
         {
             public Guid id;
             public Command(Guid id)
@@ -18,23 +19,28 @@ namespace Application.Acitivities
             }
         }
 
-        public class Handle : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
-            public Handle(DataContext context)
+            
+            public Handler(DataContext context)
             {
                 _context = context;
             }
-            async Task<Unit> IRequestHandler<Command, Unit>.Handle(Command request, CancellationToken cancellationToken)
+           
+            async Task<Result<Unit>> IRequestHandler<Command, Result<Unit>>.Handle(Command request, CancellationToken cancellationToken)
             {
-
                 Activity activity = await _context.Activities.FindAsync(request.id);
+                
+                if(activity == null) return null;
                 
                 _context.Activities.Remove(activity);
                 
-                await _context.SaveChangesAsync();
+                var res = await _context.SaveChangesAsync() > 0;
+                if(!res) return Result<Unit>.Faliure("failed to delete activity");
 
-                return Unit.Value;
+
+                return Result<Unit>.Sucess(Unit.Value);
             }
         }
